% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target-builder.R
\name{build_loader_call}
\alias{build_loader_call}
\title{Build a loader function call}
\usage{
build_loader_call(loader_fn, source)
}
\arguments{
\item{loader_fn}{Loader function name (string)}

\item{source}{Source config}
}
\value{
unevaluated function call
}
\description{
Build a loader function call
}
\keyword{internal}
