% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target-builder.R
\name{build_output_target}
\alias{build_output_target}
\title{Build an output (save) target}
\usage{
build_output_target(output)
}
\arguments{
\item{output}{Output config (list with name, format, path)}
}
\value{
tar_target() call
}
\description{
Build an output (save) target
}
\keyword{internal}
