% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target-builder.R
\name{build_source_target}
\alias{build_source_target}
\title{Build a source (loader) target}
\usage{
build_source_target(source)
}
\arguments{
\item{source}{Source config (list with name, type, path, options)}
}
\value{
tar_target() call
}
\description{
Build a source (loader) target
}
\keyword{internal}
