% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target-builder.R
\name{build_transform_target}
\alias{build_transform_target}
\title{Build a transform target}
\usage{
build_transform_target(transform)
}
\arguments{
\item{transform}{Transform config (list with name, input, function, params)}
}
\value{
tar_target() call
}
\description{
Build a transform target
}
\keyword{internal}
