% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target-builder.R
\name{get_loader_function}
\alias{get_loader_function}
\title{Get appropriate loader function name based on file type}
\usage{
get_loader_function(type)
}
\arguments{
\item{type}{File type (csv, parquet, rds)}
}
\value{
Function name as string
}
\description{
Get appropriate loader function name based on file type
}
\keyword{internal}
