% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target-builder.R
\name{get_save_function}
\alias{get_save_function}
\title{Get appropriate save function name based on format}
\usage{
get_save_function(format)
}
\arguments{
\item{format}{Output format (csv, parquet, rds)}
}
\value{
Function name as string
}
\description{
Get appropriate save function name based on format
}
\keyword{internal}
