% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{read_csv}
\alias{read_csv}
\title{Read CSV file}
\usage{
read_csv(file, ...)
}
\arguments{
\item{file}{Path to CSV file}

\item{...}{Additional arguments passed to readr::read_csv()}
}
\value{
Data frame
}
\description{
Read CSV file
}
