% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{read_parquet}
\alias{read_parquet}
\title{Read Parquet file}
\usage{
read_parquet(file, ...)
}
\arguments{
\item{file}{Path to Parquet file}

\item{...}{Additional arguments passed to arrow::read_parquet()}
}
\value{
Data frame
}
\description{
Read Parquet file
}
