% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{read_rds}
\alias{read_rds}
\title{Read RDS file}
\usage{
read_rds(file, ...)
}
\arguments{
\item{file}{Path to RDS file}

\item{...}{Additional arguments passed to readRDS()}
}
\value{
Object restored from RDS file
}
\description{
Read RDS file
}
