% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_rds}
\alias{save_rds}
\title{Save RDS file}
\usage{
save_rds(x, file, ...)
}
\arguments{
\item{x}{Object to save}

\item{file}{Output file path}

\item{...}{Additional arguments passed to saveRDS()}
}
\value{
Invisibly returns the output file path.
}
\description{
Save an object to an RDS file, creating parent directories as needed.
}
