% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_csv}
\alias{write_csv}
\title{Write CSV file}
\usage{
write_csv(x, file, ...)
}
\arguments{
\item{x}{Data frame to save}

\item{file}{Output file path}

\item{...}{Additional arguments passed to readr::write_csv()}
}
\value{
Invisibly returns x
}
\description{
Write data to a CSV file, creating parent directories as needed.
}
